Refactor API
Primarily this splits things into two APIs. The "system" API, in
the new clostrum-sys package, is "low level". It does not do error
checking or otherwise attempt to maintain consistency with Common
Lisp semantics or implement its operators directly. The clostrum
package is then the "high level" API, which is written in terms of
the low level API and does attempt to maintain consistency and
signal errors. The idea is that an implementation of environments
need only specialize the relatively simple low level API. If it
does, the high level API should then just work. This also reduces
code duplication in that environment implementations do not need
to themselves implement all of CL's complications.

Clients may specialize the high level API functions, but shouldn't
need to unless they want them to signal errors, or because they
have a different idea of the functions' semantics.

Also in this vein, the protocol classes are removed. This should
hopefully make it more possible to use Clostrum as a protocol to
existing implementation environments.

New interfaces for cells are added so that their nature is an
implementation detail.

Classes are given their own entries and cells.

Various semantics for standard operators are changed. For instance,
it is now possible to alter a symbol's value without making it
special, or interfering with its existence as a symbol macro.