New definition for ensure-FOO-cell functions
As per discussion in #sicl linked below, the idea is that if you
have a child environment you cannot mutate a parent environment.
So, when you request a cell from an environment, if it doesn't
have one already, it makes a new cell in that environment. And to
allow values to be inherited, the value from any parent cell will
be copied into that new cell.

https://irclog.tymoon.eu/libera/%23sicl?around=1698154381#1698154381