Don't erase a class with the same name as a type
The standard is not clear about how deftype and defclass interact,
or indeed how type redefinition works at all. However, it makes
some sense to have both a type expander and a class for some
standard types, such as VECTOR. So it makes sense to be a little
more flexible and allow this. Clients can enforce otherwise if
they want.