Rearrange files & delete old documentation
ASDF likes system files to be in the top directory, and I think it's
nicer to navigate without the extra level anyway, at least as long
as we have as few files as we do. I'm also hoping this fixes
source locations for the Staple documentation, but we'll see.

The old docs are out of date and I've never been able to build
them anyway. We can recover them from history if needed.